+++ resources/views/admin/pages/create.blade.php
@@ -60,7 +60,7 @@
                 </div>
                 @endif
 
-                <x-hook name="admin.create_page_main_boxes" :params="['page' => $page]" desc="Dentro da coluna principal na criação de página" />
+                <x-hook name="admin.create_page_main_boxes" desc="Dentro da coluna principal na criação de página" />
             </div>
 
             <div class="aside-column">
@@ -188,7 +188,7 @@
                     </article>
                 </div>
 
-                <x-hook name="admin.create_page_side_boxes" :params="['page' => $page]" desc="Dentro da coluna lateral na criação de página" />
+                <x-hook name="admin.create_page_side_boxes" desc="Dentro da coluna lateral na criação de página" />
             </div>
         </div>