+++ resources/views/admin/posts/create.blade.php
@@ -61,7 +61,7 @@
                 </div>
                 @endif
 
-                <x-hook name="admin.create_post_main_boxes" :params="['post' => $post]" desc="Dentro da coluna principal na criação de post" />
+                <x-hook name="admin.create_post_main_boxes" desc="Dentro da coluna principal na criação de post" />
             </div>
 
             <div class="aside-column">
@@ -192,7 +192,7 @@
                     </article>
                 </div>
 
-                <x-hook name="admin.create_post_side_boxes" :params="['post' => $post]" desc="Dentro da coluna lateral na criação de post" />
+                <x-hook name="admin.create_post_side_boxes" desc="Dentro da coluna lateral na criação de post" />
             </div>
         </div>