+++ resources/views/components/rendered/admin/menu.blade.php
@@ -1,0 +1,3 @@
+<x-hook name="admin.menu" desc="Menu na administração">
+    @include('admin.partials.menu')
+</x-hook>