+++ routes/admin.php
@@ -81,6 +81,8 @@ Route::middleware($middlewares)->prefix('admin')->name('admin.')->group(function
         Route::post('plugins/marketplace/install', [PluginMarketplaceController::class, 'installBatch'])->name('plugins.marketplace.install');
         Route::post('plugins/marketplace/refresh', [PluginMarketplaceController::class, 'refresh'])->name('plugins.marketplace.refresh');
         Route::delete('plugins/marketplace/remove/{folder}', [PluginMarketplaceController::class, 'remove'])->name('plugins.marketplace.remove');
+        Route::delete('plugins/marketplace/remove-inactive', [PluginMarketplaceController::class, 'removeInactive'])
+            ->name('plugins.marketplace.remove-inactive');
 
         // Temas
         Route::get('themes', [ThemeController::class, 'index'])->name('themes.index');
@@ -92,6 +94,8 @@ Route::middleware($middlewares)->prefix('admin')->name('admin.')->group(function
         Route::post('themes/marketplace/install', [ThemeMarketplaceController::class, 'installBatch'])->name('themes.marketplace.install');
         Route::post('themes/marketplace/refresh', [ThemeMarketplaceController::class, 'refresh'])->name('themes.marketplace.refresh');
         Route::delete('themes/marketplace/remove/{folder}', [ThemeMarketplaceController::class, 'remove'])->name('themes.marketplace.remove');
+        Route::delete('themes/marketplace/remove-inactive', [PluginMarketplaceController::class, 'removeInactive'])
+            ->name('themes.marketplace.remove-inactive');
     });
 
     // Dashboard